fix: extracted mui types, renaming and cleanup
